Backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
celery[redis]
redis
django-environ
python-dotenv
django-cors-headers


